test: Participants 테스트 추가
