+++ src/test/java/domain/BettingTableTest.java
@@ -1,0 +1,8 @@
+package domain;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class BettingTableTest {
+
+
+}
